# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppAgendaApi",
        "description": "Lista de Endspoints",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1/"
        }
    ],
    "paths": {
...
        "/favorite": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Obtiene la lista de favoritos del usuario autenticado",
                "operationId": "40cccb9d38c80baf7f3ba8dc2b5bb2aa",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtro por nombre del contacto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Filtro por número de teléfono del contacto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "Filtro por apodo del contacto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de favoritos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FavoriteResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener la lista de favoritos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error: [detalles del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Store a new favorite contact",
                "operationId": "b61cce84e018fc68a7cab2f6b4e4f236",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FavoriteRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favorite contact successfully stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha creado añadido exitosamente el contacto a favoritos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El recurso solicitado no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to add this contact to favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No estás autorizado para añadir este contacto a favoritos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error: [detalles del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/favorite/{id}": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Show a favorite contact",
                "operationId": "416e3981c88ac49b4ad7dc14c1ee95cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the favorite contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite contact found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Favorite contact not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se ha encontrado el contacto favorito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to view this favorite contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No estás autorizado para ver este contacto favorito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error: [detalles del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorite/{id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Delete a favorite contact",
                "operationId": "1a83276ced9917dd0a6dca0470204c50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the favorite contact to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite contact deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contacto favorito Eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Favorite contact not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se ha encontrado el contacto favorito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to delete this favorite contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No estás autorizado para elimiinar este contacto favorito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error: [detalles del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FavoriteRegisterRequest": {
                "title": "Favorite Register Request",
                "required": [
                    "contact_id"
                ],
                "properties": {
                    "contact_id": {
                        "type": "string",
                        "example": "encrypted_contact_id_here"
                    }
                },
                "type": "object"
            },
...
            "FavoriteResource": {
                "title": "Favorite Resource",
                "description": "Favorite resource representation",
                "properties": {
                    "id": {
                        "description": "ID of the favorite (encrypted)",
                        "type": "string",
                        "example": "eyJpdiI6IjZkZ3BMRnl2Nlducmd4RWlCNDAxZkE9PSIsInZhbHVlIjoianpyOEorMWdwcmlRaWJPcUtlcFd0QT09IiwibWFjIjoiNTE2MTkxZDA0YTgyNTI1OTkwN2IxMjJlMmM1ZjNhZTQwYjNlYjIxOTBhNTI2NDcwMDM2YjkwYWY0YjVmNzYwOCIsInRhZyI6IiJ9"
                    },
                    "contact": {
                        "title": "Contact details",
                        "description": "Details of the contact that was favorited",
                        "properties": {
                            "id": {
                                "description": "ID of the contact (encrypted)",
                                "type": "string",
                                "example": "eyJpdiI6Im53enBpS0xGZjF6VFUyaG1aQ2E3UFE9PSIsInZhbHVlIjoiaVJDdHZrZitLZ09HQ2YxMm84OUYvUT09IiwibWFjIjoiYmM2YzdlNjI1MjA4NjViOWJiYmIwMDBhN2U3ZjYyOGFmNzM3ZGY4YzJkZDI2ZjViYTY1NDkxNGIzY2I2YTM2YyIsInRhZyI6IiJ9"
                            },
                            "name": {
                                "description": "Name of the contact",
                                "type": "string",
                                "example": "Juanito"
                            },
                            "phone": {
                                "description": "Phone number of the contact",
                                "type": "string",
                                "example": "0993954831"
                            },
                            "nickname": {
                                "description": "Nickname of the contact",
                                "type": "string",
                                "example": null
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost/api/contacts"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost/api/contacts?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost/api/contacts?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost/api/contacts?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación",
                "name": "BearerToken",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}